Refactor folder structure to enhance code readability and maintainability.
